Split commands/history.rs into target and related submodules
history/mod.rs keeps the history/related/refresh entry points and
project resolution; target.rs resolves which project repo and pathspecs
a related query is about; related.rs owns the git-to-Knit-history join
and result rendering. Pure moves, no behavior change.

Knit-Group: kg_20260612_664482
Knit-Bundle: open-source-prep
